requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
rapidfuzz>=3.0.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
